implemented more

+++ src/main/kotlin/org/three/math/Plane.kt
@@ -5,12 +5,12 @@ package org.three.math
 open external class Plane {
 
     constructor()
-    constructor(normal: Vector3, constant: Double)
+    constructor(normal: Vector3, constant: Float)
 
     var normal: Vector3
-    var constant: Double
+    var constant: Float
 
-    fun set(normal: Vector3, constant: Double)
+    fun set(normal: Vector3, constant: Float)
 
     fun clone() : Plane
     fun copy(plane: Plane) : Plane
