+++ src/main/kotlin/org/three/math/Vector4.kt
@@ -5,11 +5,11 @@ package org.three.math
 open external class Vector4 {
 
     constructor()
-    constructor(x: Double, y: Double, z: Double, w: Double)
+    constructor(x: Float, y: Float, z: Float, w: Float)
 
-    var x: Double
-    var y: Double
-    var z: Double
-    var w: Double
+    var x: Float
+    var y: Float
+    var z: Float
+    var w: Float
 
 }
